Dro ut Periode og FormKomponent i egne filer, og skrev alt om til typescript.
